added option to remove potentially risky T1w and T2w files

+++ delete_identifiable_DICOMs.py
@@ -1,3 +1,10 @@
+"""
+Delete DICOM files that could lead to identifying the patient.
+"""
+
+__author__ = "Jonas Richiardi"
+__credits__ = ["Jonathan Patiño-Lopez", "Adrien Jammot", "Veronica Ravano"]
+__status__ = "development"
 
 import argparse
 import os
@@ -11,15 +18,20 @@ import pydicom
 from tqdm import tqdm
 
 IMAGETYPES_TO_REMOVE=['SCREEN SAVE','DISPLAY','LOCALIZER','OTHER']
+T1W_TO_REMOVE=['tfl3d','fl2d']
 
 
 def delete_identifiable_dicom_file(
-        filename: str) -> bool:
+        filename: str,
+        delete_T1w: bool = False,
+        delete_T2w: bool = False) -> bool:
     """
     If identifiable data is present, deletes the Dicom file
 
     Args :
         filename: path to dicom image.
+        delete_T1w: also delete potentially identifiable (face-reconstructible) T1w images like MPRAGEs
+        delete_T2w: also delete potentially identifiable (face-reconstructible) T2w images like FLAIRs
 
     Returns:
         bool: whether the file was deleted (True) or not (False)
@@ -50,7 +62,7 @@ def delete_identifiable_dicom_file(
             delete_this_file = True
 
     if not delete_this_file and ("ProtocolName" in attributes):
-        my_re_pn=re.compile('(?i).*(Scout|localizer|t2_haste_sag_ipat2).*')
+        my_re_pn=re.compile('(?i).*(Scout|localizer).*')
         if my_re_pn.search(dataset.data_element('ProtocolName').value) is not None:
             delete_this_file = True
 
@@ -74,6 +86,23 @@ def delete_identifiable_dicom_file(
         if my_re_sd_Key.search(dataset.data_element('SeriesDescription').value) is not None:
             delete_this_file = True
 
+    if not delete_this_file and delete_T1w:
+        # Sagittal 2D FLASH (Vida): SequenceName *fl2d1, ScanningSequence: GR, ImageType ORIGINAL\PRIMARY
+        # mprage: ImageType ORIGINAL\PRIMARY, sequenceName tfl3d
+        if ("SequenceName" in attributes) and ("ImageType" in attributes):
+            if any([this_seqname in dataset.data_element('SequenceName').value for this_seqname in T1W_TO_REMOVE]) and 'ORIGINAL' in dataset.data_element('ImageType') :
+                delete_this_file = True
+
+        raise NotImplementedError('delete_T1w is not implemented')
+
+    if not delete_this_file and delete_T2w:
+        # Transverse 2D FLAIR (turbo inversion recovery): SequenceName *tir2d1_15, ScanningSequence: SE, MRAcquisitionType: 2D, ImageType ORIGINAL\PRIMARY
+        # Other FLAIR: SequenceName: spcir, MRAcquisitionType: 3D
+        # -> SequenceName: .*'?IR'?.* not case sensitive
+        if ("SequenceName" in attributes) and ("ImageType" in attributes) :
+            if 'ir' in dataset.data_element('SequenceName').value and 'ORIGINAL' in dataset.data_element('ImageType'):
+                delete_this_file = True
+
     if delete_this_file:
         os.remove(filename)
 
@@ -82,7 +111,9 @@ def delete_identifiable_dicom_file(
 
 def sanitize_all_dicoms_within_root_folder(
         datapath: str = os.path.join(".", "data"),
-        pattern_dicom_files: str = os.path.join("ses-*", "*", "*")
+        pattern_dicom_files: str = os.path.join("ses-*", "*", "*"),
+        delete_T1w: bool = False,
+        delete_T2w: bool = False
     ) -> int:
     """
     Sanitizes all Dicom images located at the datapath in the structure specified by pattern_dicom_files parameter.
@@ -90,6 +121,8 @@ def sanitize_all_dicoms_within_root_folder(
         datapath: The path to the dicom images.
         pattern_dicom_files: the (generic) path to the dicom images starting from the patient folder. In a PACSMAN dump, this would reflect e.g.
             ses-20170115/0002-MPRAGE/*.dcm
+        delete_T1w: delete T1-weighted images that could be used to identify the patients
+        delete_T2w: delete T2-weighted images that could be used to identify the patients
     Returns :
         int : always 0
     """
@@ -127,7 +160,9 @@ def sanitize_all_dicoms_within_root_folder(
                     # Loop over all dicom files within a series and remove offending files
                     n_deleted_files_in_series = 0
                     for filename in all_filenames_series:
-                        file_deleted = delete_identifiable_dicom_file(filename)
+                        # TODO speedup - if we flag one file, we can assume the whole series can be deleted and we can
+                        #  just delete the rest of the dir
+                        file_deleted = delete_identifiable_dicom_file(filename,delete_T1w,delete_T2w)
                         if file_deleted:
                             n_deleted_files_in_series += 1
 
@@ -139,17 +174,25 @@ def main(argv):
 
     parser.add_argument("--in_folder", "-d", help="Root dir to the dicom files to be screened for identifiables files.",
                         default=os.path.join(".", "data"), required=True)
+    parser.add_argument("--delete_T1w", "-i", help="Delete potentially identifiable T1-weighted images such as MPRAGE",
+                        default=False, required=False, action='store_true')
+    parser.add_argument("--delete_T2w", "-i", help="Delete potentially identifiable T1-weighted images such as FLAIR",
+                        default=False, required=False, action='store_true')
 
     args = parser.parse_args()
 
     data_path = os.path.normcase(os.path.abspath(args.in_folder))
+    delete_T1w = args.delete_T1w
+    delete_T2w = args.delete_T2w
 
     if not os.path.isdir(data_path):
         raise NotADirectoryError('Input directory does not exist.')
 
     print("Deleting potentially identifiable Dicom files within path {}".format(os.path.abspath(data_path)))
     # Sanitize all files.
-    result = sanitize_all_dicoms_within_root_folder(datapath=data_path)
+    result = sanitize_all_dicoms_within_root_folder(datapath=data_path,
+                                                    delete_T1w=delete_T1w,
+                                                    delete_T2w=delete_T2w)
 
 
 if __name__ == "__main__":
